Add files via upload
This is the ultimate organizer. You can have up to twelve categories for the data. You are able to add and delete records and categories, as well as search for someone's record and update it if you want. You can paste your data directly into the file and it will be ready to use!